.info/requirements.txt
markdown2
PyYAML